Added new metric_battery_cycle feature (#110)
* Add new feature to account of the cost of cycling your battery
* Fix charge and discharge power settings to avoid rounding errors.

The setting input_number.metric_battery_cycle can be used to put a cost on using your battery for charging and discharging. 
In theory if you think your battery will last say 6000 complete cycles and cost you £4000 and is 9.5kWh then each cycle is 19kWh and so the cost is £4000 / 19 / 6000 = 3.5p. 
If you configure this number higher then more expensive plans will be selected which avoid charging and discharging your battery as much. 
The default is 0 which means ignore the cycle costs in the plan. Note that the cycle cost will not be included in the cost predictions, just taken into account in the planning stage.