Fix stale inverter data
After the introduction of 74a46fa6c80c925c866ac2b9f36151a569faa556 the inverter data
when using the ge_cloud native addon becomes stale and doesn't update very frequently.
This change is to force the fetch of the inverter data regardless of what we do with the
plan.

related #2221