Update the configuration guide to reference rate overrides
As standard, Predbatt will schedule cycles of charge and discharge of the battery using the export period, as this is the most economic option (7p import / 15p export will make it economically feasible despite charge losses).

It's (IMHO) not in the spirit of the tariff, and Octopus are "not obliged" to pay for brown exports under 4.5.1(i) of their export terms.  IE Octopus could do something about it if enough people are operating outside of the spirit of the terms. 

Took me a while to work out how to work around this, thought it would help others.