Correct Conversion of SoC Percent to SoC kWh for Charge Curve
When a SoC was in %, the code needs to convert it to kWh. However it was using a function that converts kWh->%, which meant the values were nonsense.

Add a new utility that does the same conversion, but in the opposite direction, %->kWh, and apply it in the charge curve code.