Allow charge curves to detect multi-% jumps
With some batteries it is possible for the final charge curve to jump more than one percent within any given minute, particularly towards the end of the curve.

Allow the code to correctly handle this by averaging the power across the total number of percentage points crossed in a given step. For example if SoC charged 2% in 1 minute, then the average charger each of the two %age points will actually be half the calculated amount.