Fix (Dis)Charge Curve for Inverters with `soc_percent` (#2986)
* Correct Conversion of SoC Percent to SoC kWh for Charge Curve

When a SoC was in %, the code needs to convert it to kWh. However it was using a function that converts kWh->%, which meant the values were nonsense.

Add a new utility that does the same conversion, but in the opposite direction, %->kWh, and apply it in the charge curve code.

* Correct Charge Curve Logging Order

Logs are shown in reverse. The actual values were correctly reversed, but the heading was still shown at the bottom. By adding at the end, it will display in the logs in the correct order to copy to YAML.

* Allow charge curves to detect multi-% jumps

With some batteries it is possible for the final charge curve to jump more than one percent within any given minute, particularly towards the end of the curve.

Allow the code to correctly handle this by averaging the power across the total number of percentage points crossed in a given step. For example if SoC charged 2% in 1 minute, then the average charger each of the two %age points will actually be half the calculated amount.

* Remove Change to Charge Curve Average Power

The change should not have modified the average power, but is partially still required to ensure multi-% changes are detected.

* Remove Superfluous change to Utils.py

The changes were made in a different way to main, so this is not required...

* Revert "Correct Charge Curve Logging Order"

This reverts commit 061e075f4ea8305652f554d608f54337af4e1c09.