Remove minimum drop for reset to zero
If it is incrementing data, then any drop is technically invalid and should be ignored.

By requiring a minimum drop of 1, then for cases with a unit of kWh, then on days where the entity (e.g. PV production, or export energy) has less than 1kWh overall for the day, then the resulting data is wrong by a substantial amount. See #2925.