Undo change to SIG `has_target_soc` and `has_reserve_soc` (#3124)
Commit aa180dce9d42131db444b11d22d03765d76ce60a altered the config.py
values for `has_target_soc` and `has_reserve_soc` to True in `config.py`,
referencing issue #2077. However, this causes Predbat to set the target SOC
to 0% (or some other minimum reserve value) when in Demand and Freeze
Charge mode, which forbids the system from charging the battery from solar.
Looking at the discussion in #2077, it appears that this change was stated
as not working correctly by its original proposer, and all responses have
concurred that it is not working correctly.

Revert the change and set the parameters back to False.